+++ app/windows/flutter/generated_plugins.cmake
@@ -3,11 +3,8 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  media_kit_libs_windows_video
-  media_kit_video
-  syncfusion_pdfviewer_windows
+  app_links
   url_launcher_windows
-  volume_controller
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST